Limit length, add handling for invalid calculations
If the result of a calculation is too large, "TOO LARGE" is displayed.
If the user tries to append digits beyond the length or decimal place limit, input is blocked.
If a division by 0 is performed, "ERROR" is displayed.
If "TOO LARGE" or "ERROR" is used in a calculation, the result becomes "ERROR".